fix(host): drop redundant u32 casts on FreeBSD kinfo uid/gid
kinfo_proc.ki_ruid/ki_rgid are already u32, so `as u32` is a same-type
cast that clippy's `unnecessary_cast` flags — failing the FreeBSD clippy
lane (`-D warnings`). ki_ppid/ki_pgid stay `i32 -> u32`. Pre-existing;
surfaced now that the BSD backend is clippy-gated in CI.

+++ crates/carrick-host/src/host_proc.rs
@@ -641,8 +641,8 @@ mod imp {
             state: freebsd::state_char(kp.ki_stat),
             ppid: kp.ki_ppid.max(0) as u32,
             pgid: kp.ki_pgid.max(0) as u32,
-            uid: kp.ki_ruid as u32,
-            gid: kp.ki_rgid as u32,
+            uid: kp.ki_ruid,
+            gid: kp.ki_rgid,
             comm: freebsd::comm(&kp),
         })
     }
